@@ -1,10 +1,10 @@
 package com.alex.multimoduleproject
 
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.alibaba.android.arouter.facade.annotation.Route
 
-@Route(path = "/app/Main2Activity")
+@Route(path = "/app/activity/Main2Activity")
 class Main2Activity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
