@@ -1,7 +1,7 @@
 package com.alex.module1
 
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.alibaba.android.arouter.facade.annotation.Route
 
 @Route(path = "/module1/Module1Activity")
