@@ -1,10 +1,10 @@
 package com.alex.module2
 
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import com.alibaba.android.arouter.facade.annotation.Route
 
-@Route(path = "/module2/Module2Activity")
+@Route(path = "/module2/activity/Module2Activity")
 class Module2Activity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
